+++ src/material/slide-toggle/testing/BUILD.bazel
@@ -21,27 +21,7 @@ ts_library(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/material/slide-toggle",
-    ],
-)
-
-ng_web_test_suite(
-    name = "unit_tests",
-    deps = [
-        ":unit_tests_lib",
-    ],
-)
-
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/testing",
@@ -49,5 +29,12 @@ ng_test_library(
         "//src/material/slide-toggle",
         "@npm//@angular/forms",
         "@npm//@angular/platform-browser",
+    ],
+)
+
+ng_web_test_suite(
+    name = "unit_tests",
+    deps = [
+        ":unit_tests_lib",
     ],
 )
